Move SetupScreen folder to parent folder
