+++ lib/main.dart
@@ -4,11 +4,12 @@ import 'package:flutter_my_bakery/screens/authenticate/sign_in.dart';
 import 'package:flutter_my_bakery/screens/home/bottom_bar_state.dart';
 import 'package:firebase_core/firebase_core.dart';
 
-Future<void> main() async{
+Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
   runApp(MyApp());
 }
+
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -22,16 +23,17 @@ class MyApp extends StatelessWidget {
     );
   }
 }
+
 class MainScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return StreamBuilder<User>(
-      stream: FirebaseAuth.instance.authStateChanges(),
-      builder: (context, snapshot){
-      if(snapshot.hasData && snapshot.data != null){
-        return BottomBarState();
-      }
-        return SignIn();
-      });
+        stream: FirebaseAuth.instance.authStateChanges(),
+        builder: (context, snapshot) {
+          if (snapshot.hasData && snapshot.data != null) {
+            return BottomBarState();
+          }
+          return SignIn();
+        });
   }
 }
