+++ lib/models/Product.dart
