+++ lib/shared/states.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_my_bakery/SetupScreen/screens/firstPage.dart';
+import 'package:flutter_my_bakery/screens/setup/firstPage.dart';
 import 'package:flutter_my_bakery/screens/home/home.dart';
 import 'package:flutter_my_bakery/screens/home/my_businesses.dart';
